% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend.R
\name{extend}
\alias{extend}
\title{Extend a base Singularity container to add suites lacking container support}
\usage{
extend(staticMdiDir)
}
\description{
This function is only intended to be called by a Singularity instance.
It is not for general use and should be ignored by most people.
}
\details{
Packages not present in container are installed into an active, bind-mounted
MDI installation.
}
