% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keys.R
\name{password_store}
\alias{password_store}
\title{Generate a salted and hashed access key suitable for storing on disk}
\usage{
password_store(key)
}
\arguments{
\item{key}{character. The access key that will by typed into the web
interface by a user to gain access to your MDI apps server.}
}
\value{
A hash value ready for pasting into config file 'stage2-apps.yml'.
}
\description{
Hash values generated by \code{mdi::password_store()} are the strings that
should be pasted into config file 'stage2-apps.yml'.
}
\details{
Please note that access_control 'keys' offer only modest security.
Their main utility is to prevent people from using your server more than
to enforce strict data protection. You should protect your server through 
additional means such as an IP firewall and ensure that your MDI installation 
files are secure.
}
