% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\alias{develop}
\alias{ondemand}
\title{Run the Michigan Data Interface (MDI) in a web server}
\usage{
run(
  rootDir = "~",
  dataDir = NULL,
  ondemandDir = NULL,
  url = "http://localhost:3838/",
  mode = "local",
  browser = mode != "server",
  debug = mode != "server",
  developer = FALSE,
  checkout = NULL
)

develop(rootDir = "~", dataDir = NULL, url = "http://localhost:3838/")

ondemand(ondemandDir, rootDir = "~", dataDir = NULL, port = 3838)
}
\arguments{
\item{rootDir}{character. Path to the directory where the MDI
has previously been installed. Defaults to your home directory.}

\item{dataDir}{character. Path to the directory where your MDI
data can be found. Defaults to code{rootDir}/data. You might wish to change
this to a directory that holds shared data, e.g., for your laboratory.}

\item{ondemandDir}{character. Path to the directory where a public
installation of the MDI can be found. Some paths from this
installation will be used instead of the user's installation when code{mode}
is 'ondemand'.}

\item{url}{character. The complete browser URL to load the web page. Must
include port 3838 and a trailing slash. Examples: 'http://localhost:3838/' 
or 'https://mymdi.org:3838/'.}

\item{mode}{character. Either 'server', 'local' or 'ondemand'. Most users
want 'local' (the default) to run the MDI on your desktop or laptop. 
Mode 'server' is for a public web server, 'ondemand' is for managed execution
on an HPC cluster via Open OnDemand Interactive Apps.}

\item{browser}{logical. Whether or not to attempt to launch a web browser 
after starting the MDI server. Defaults to TRUE unless \code{mode} is 'server'.}

\item{debug}{logical. When \code{debug} is TRUE, verbose activity logs
will be printed to the R console where \code{mdi::run} was called.
Defaults to TRUE unless \code{mode} is 'server'.}

\item{developer}{logical. When \code{developer} is TRUE, additional
development utilities are added to the web page. Only honored if
\code{mode} is 'local'.}

\item{checkout}{character. If NULL (the default), \code{run} will auto-set
repositories to the appropriate branch based on R and MDI versions.
Developers might want to specify a code branch.}
}
\value{
These functions never return. They launch a blocking web server 
that runs perpeptually in the parent R process in the global environment.
}
\description{
\code{run} launches the suites of data analysis applications that comprise
the Michigan Data Interface (MDI) in a web server, either on a web host 
that is publicly addressable, on your local computer, or on on a cluster 
compute node within an OnDemand batch job. \code{develop} is a shortcut 
to \code{run}  with settings appropriate for developers (mode='local', 
browser=FALSE, debug=TRUE, developer=TRUE).
}
\details{
All default settings are consistent with an end user running the 
MDI in local mode on their desktop or laptop computer.

\code{rootDir} must be the same directory as used in a prior
call to \code{mdi::install}.

When \code{developer} is TRUE, you must have git properly installed on
your computer. If needed, you will be prompted to provide your email
address and user name on first use, which will be stored in your local
repositories and used to tag your code commits. To pull or push code
via the GUI, you must also have enabled non-prompted authorized
access to the remote repositories (e.g. via the command line).

When \code{developer} is TRUE, \code{run()} always ensures that the current
git branch is somewhere other than 'main' (either 'develop' or your
current development branch). Otherwise, it resets the git branch to 'main'
at the appropriate version tag based on your system and usage.
}
